+++ resources/2.3/ma3_dummy_object_free_no_doc.lua
@@ -253,7 +253,7 @@ end
 function SelectionNotifyObject(objectToNotify)
 end
 
----@return table subFixtureIndexes
+---@return number[] subFixtureIndexes
 function SelectionTable()
     return {}
 end